Inlineable reader.Seek and no-lock return on invalid whence
Ugly solution, but couldn't find an elegant way of checking for a bad whence without checking valid cases first. Could use a closure to reduce code duplication, but that seems like overkill for only a few cases.
Note that returning 0 on an invalid whence is the behavior of `strings.Reader` and `bytes.Reader`.